requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
pytesseract==0.3.10
ddddocr==1.4.11
gmssl==3.2.2
requests-toolbelt==1.0.0